+++ dev/src/js/src/physics/lite/manager.js
@@ -40,10 +40,11 @@ var Lightning;
             if (this._paused)
                 return;
             for (var i in this._pools) {
+                this.checkPoolCollisions(this._pools[i]);
                 for (var _i = 0, _a = this._pools[i].bodies; _i < _a.length; _i++) {
                     var body = _a[_i];
-                    this.outOfBounds(body);
-                    // this.checkWorldCollide(body);
+                    // this.outOfBounds(body);
+                    this.checkWorldCollide(body);
                     this.updatePosition(body);
                     body.objRef.updateTransform();
                 }
@@ -84,25 +85,62 @@ var Lightning;
             //left
             if (body.x <= this._worldBounds.x) {
                 body.velocity.x *= -1;
-                console.log('right');
+                console.log('left');
             }
             //right
-            if (body.x >= this._worldBounds.x + this._worldBounds.width) {
+            if (body.x >= this._worldBounds.width) {
                 body.velocity.x *= -1;
             }
             //down
-            if (body.y <= this._worldBounds.y) {
+            if (body.y <= this._worldBounds.height) {
                 body.velocity.y *= -1;
             }
             //up
-            if (body.y >= this._worldBounds.y + this._worldBounds.height) {
+            if (body.y >= this._worldBounds.y) {
                 body.velocity.y *= -1;
             }
         };
+        LitePhysicsManager.prototype.checkPoolCollisions = function (pool) {
+            if (pool.bodies.length < 2)
+                return;
+            var c = 1;
+            for (var _i = 0, _a = pool.bodies; _i < _a.length; _i++) {
+                var body = _a[_i];
+                for (var i = c; i < pool.bodies.length; i++) {
+                    // body = the body we're currently on
+                    // the body we're checking against
+                    var body2 = pool.bodies[i];
+                    // check two colisions here
+                    if (body.x < body2.x + body2.bounds.width &&
+                        body.x + body.bounds.width > body2.x &&
+                        body.y < body2.y + body2.bounds.height &&
+                        body.bounds.height + body.y > body2.y) {
+                        body.velocity.x *= -1;
+                        body.velocity.y *= -1;
+                        body2.velocity.x *= -1;
+                        body2.velocity.y *= -1;
+                    }
+                }
+                c++;
+            }
+        };
         LitePhysicsManager.prototype.outOfBounds = function (body) {
-            if (body.x <= this._worldBounds.x + body.bounds.width) {
-                console.log('out of bounds');
-                // console.log(body.bounds.width);
+            console.log(body.x, this._worldBounds.x);
+            //left
+            if (body.x <= this._worldBounds.x) {
+                console.log('out of bounds left');
+            }
+            //right
+            if (body.x >= this._worldBounds.width) {
+                console.log('out of bounds right');
+            }
+            //down
+            if (body.y >= this._worldBounds.height) {
+                console.log('out of bounds down');
+            }
+            //up
+            if (body.y <= this._worldBounds.y) {
+                console.log('out of bounds up');
             }
         };
         return LitePhysicsManager;