+++ dev/src/ts/states/game.ts
@@ -8,43 +8,48 @@ export default class GameState extends Lightning.State {
      private _start:boolean = false;
 
     create() {
+
         this.game.physics.lite.enablePhysics();
-        this._sprite = new Lightning.Sprite();
-        let texture:Lightning.Texture =  Lightning.Geometry.Rect(20, 20).generateTexture();
-        this._sprite.texture = texture;
-        this._sprite.y = this.game.height / 2;
-        this._sprite.x += 50;
-        this._sprite.setAnchor(0.5);
-        this.add(this._sprite);
-
+        let texture:Lightning.Texture =  Lightning.Geometry.Rect(5, 5).generateCanvasTexture();
         let pool = this.game.physics.lite.createPool('test');
-        this._sprite.enablePhysicsBody();
-        pool.add(this._sprite.body);
-        this._sprite.body.velocity.x = 2;
-        this._sprite.body.enableDebug();
+
+        for(var i = 0; i < 1; i++) {
+            let sprite = new Lightning.Sprite();
+            sprite.texture = texture;
+            sprite.x = Lightning.Maths.rngFloat(0, this.game.width);            
+            sprite.y = Lightning.Maths.rngFloat(0, this.game.height);
+            // sprite.rotation = Lightning.Maths.rngFloat(0, 12);
+            this.add(sprite);
+
+            sprite.enablePhysicsBody();
+            pool.add(sprite.body);
+            sprite.body.enableDebug();
+
+            sprite.body.velocity.x = Lightning.Maths.rngFloat(-5, 5);
+            sprite.body.velocity.y = Lightning.Maths.rngFloat(-5, 5);
+        }
+
+        let bigGuy = new Lightning.Sprite();
+        bigGuy.texture = Lightning.Geometry.Rect(50, 50).generateCanvasTexture();
+        bigGuy.x = this.game.width / 2;
+        bigGuy.y = this.game.height / 2;
+        this.add(bigGuy);
+
+        bigGuy.enablePhysicsBody();
+        bigGuy.body.enableDebug();
+
+        let cEvent = this.game.physics.lite.createCollisionEvent('t', bigGuy.body, pool.bodies);
+        cEvent.onCollide((obj1:Lightning.LitePhysicsBody, obj2:Lightning.LitePhysicsBody) => {
+            obj2.velocity.x *= -1;
+            obj2.velocity.y *= -1;
+            obj2.pauseCollisionDetection = true;
+
+            setTimeout(() => {
+                obj2.pauseCollisionDetection = false;
+            }, 20);
+        });
+
         
-
-
-        // setTimeout(() => {
-        //     sprite.enablePhysicsBody();
-        //     let pool = this.game.physics.lite.createPool('test');
-        //     pool.add(sprite.physicsBody);
-        //     sprite.physicsBody.velocity.x = 20;
-
-        //     console.log(sprite.width)
-        // }, 500);
-
-        
-
-
-        // let sprite:Lightning.Sprite = new Lightning.Sprite();
-        // let texture:Lightning.Graphics = Lightning.Geometry.Triangle(50);
-        // sprite.texture = this.game.generateTexture(texture);
-        // sprite.x = this.game.width / 2;
-        // sprite.y = this.game.height / 2;
-        // sprite.setAnchor(0.5);
-        // this.add(sprite);
-
         /**
         1.  * creating basic tween
          */