+++ src/core/events/event.ts
@@ -55,7 +55,7 @@ namespace Lightning {
                 let subscription:iEventSubscription = this._subscribers[i];
 
                 // call the stored function within the specific subscription instance
-                subscription.fn.call(subscription.ctx, subscription.params, this, params);
+                subscription.fn.call(subscription.ctx, params, subscription.params, this);
 
                 // if the subscription was added once, then remove it now
                 if(subscription.once) {